restructure, refactor the HKSourceRevision extension builder
