+++ Sources/HealthKitOnFHIR/FHIR Extensions/FHIR Extension Builder/FHIRExtensionBuilder+Metadata.swift
@@ -6,28 +6,12 @@
 // SPDX-License-Identifier: MIT
 //
 
-// swiftlint:disable file_length
-
 import Foundation
 import HealthKit
 import ModelsR4
 
 
 extension FHIRExtensionUrls {
-    // SAFETY: this is in fact safe, since the FHIRPrimitive's `extension` property is empty.
-    // As a result, the actual instance doesn't contain any mutable state, and since this is a let,
-    // it also never can be mutated to contain any.
-    /// Url of a FHIR Extension containing, if applicable, encoded `HKDevice` of the `HKObject` from which a FHIR `Observation` was created.
-    public nonisolated(unsafe) static let sourceDevice = "https://bdh.stanford.edu/fhir/defs/sourceDevice".asFHIRURIPrimitive()!
-    // swiftlint:disable:previous force_unwrapping
-    
-    // SAFETY: this is in fact safe, since the FHIRPrimitive's `extension` property is empty.
-    // As a result, the actual instance doesn't contain any mutable state, and since this is a let,
-    // it also never can be mutated to contain any.
-    /// Url of a FHIR Extension containing, if applicable, encoded `HKSourceRevision` of the `HKObject` from which a FHIR `Observation` was created.
-    public nonisolated(unsafe) static let sourceRevision = "https://bdh.stanford.edu/fhir/defs/sourceRevision".asFHIRURIPrimitive()!
-    // swiftlint:disable:previous force_unwrapping
-    
     // SAFETY: this is in fact safe, since the FHIRPrimitive's `extension` property is empty.
     // As a result, the actual instance doesn't contain any mutable state, and since this is a let,
     // it also never can be mutated to contain any.
@@ -38,91 +22,10 @@ extension FHIRExtensionUrls {
 
 
 extension FHIRExtensionBuilderProtocol where Self == FHIRExtensionBuilder<HKObject> {
-    /// A FHIR Extension Builder that writes encoded `HKDevice` of a HealthKit sample into a FHIR `Observation` created from the sample.
-    public static var sourceDevice: FHIRExtensionBuilder<HKSample> {
-        .init { (sample: HKSample, observation) in
-            guard let device = sample.device else {
-                observation.removeAllExtensions(withUrl: FHIRExtensionUrls.sourceDevice)
-                return
-            }
-            let deviceInfo = Extension(url: FHIRExtensionUrls.sourceDevice)
-            let appendDeviceInfoEntry = { (keyPath: KeyPath<HKDevice, String?>) in
-                guard let name = keyPath._kvcKeyPathString else {
-                    print("Unable to obtain name for keyPath '\(keyPath)'. Skipping.")
-                    return
-                }
-                guard let value = device[keyPath: keyPath] else {
-                    return
-                }
-                let url = FHIRExtensionUrls.sourceDevice.appending(component: name)
-                deviceInfo.appendExtension(
-                    Extension(url: url, value: .string(value.asFHIRStringPrimitive())),
-                    replaceAllExistingWithSameUrl: true
-                )
-            }
-            appendDeviceInfoEntry(\.name)
-            appendDeviceInfoEntry(\.manufacturer)
-            appendDeviceInfoEntry(\.model)
-            appendDeviceInfoEntry(\.hardwareVersion)
-            appendDeviceInfoEntry(\.firmwareVersion)
-            appendDeviceInfoEntry(\.softwareVersion)
-            appendDeviceInfoEntry(\.localIdentifier)
-            appendDeviceInfoEntry(\.udiDeviceIdentifier)
-            observation.appendExtension(deviceInfo, replaceAllExistingWithSameUrl: true)
-        }
-    }
-    
-    
-    /// A FHIR Extension Builder that writes encoded `HKSourceRevision` of a HealthKit sample into a FHIR `Observation` created from the sample.
-    public static var sourceRevision: FHIRExtensionBuilder<HKSample> {
-        .init { (sample: HKSample, observation) throws in // swiftlint:disable:this closure_body_length
-            let revision = sample.sourceRevision
-            let deviceInfo = Extension(url: FHIRExtensionUrls.sourceRevision)
-            let fieldUrl = { (components: String...) in
-                FHIRExtensionUrls.sourceRevision.appending(components: components)
-            }
-            let appendDeviceInfoEntry = { (keyPath: KeyPath<HKSourceRevision, String?>) in
-                guard let name = keyPath._kvcKeyPathString else {
-                    print("Unable to obtain name for keyPath '\(keyPath)'. Skipping.")
-                    return
-                }
-                guard let value = revision[keyPath: keyPath] else {
-                    return
-                }
-                let url = fieldUrl(name)
-                deviceInfo.appendExtension(
-                    Extension(url: url, value: .string(value.asFHIRStringPrimitive())),
-                    replaceAllExistingWithSameUrl: true
-                )
-            }
-            deviceInfo.appendExtension(
-                Extension(
-                    extension: [
-                        Extension(
-                            url: fieldUrl("source", "name"),
-                            value: .string(revision.source.name.asFHIRStringPrimitive())
-                        ),
-                        Extension(
-                            url: fieldUrl("source", "bundleIdentifier"),
-                            value: .string(revision.source.bundleIdentifier.asFHIRStringPrimitive())
-                        )
-                    ],
-                    url: fieldUrl("source")
-                ),
-                replaceAllExistingWithSameUrl: true
-            )
-            appendDeviceInfoEntry(\.version)
-            appendDeviceInfoEntry(\.productType)
-            appendDeviceInfoEntry(\.OSVersion)
-            observation.appendExtension(deviceInfo, replaceAllExistingWithSameUrl: true)
-        }
-    }
-    
-    
     /// A FHIR Extension Builder that writes encoded metadata of a HealthKit sample into a FHIR `Observation` created from the sample.
-    public static var metadata: FHIRExtensionBuilder<HKSample> {
-        .init { (sample: HKSample, observation) in // swiftlint:disable:this closure_body_length
-            guard let metadata = sample.metadata, !metadata.isEmpty else {
+    public static var metadata: FHIRExtensionBuilder<HKObject> {
+        .init { (object: HKObject, observation) in // swiftlint:disable:this closure_body_length
+            guard let metadata = object.metadata, !metadata.isEmpty else {
                 observation.removeAllExtensions(withUrl: FHIRExtensionUrls.metadata)
                 return
             }
@@ -147,7 +50,7 @@ extension FHIRExtensionBuilderProtocol where Self == FHIRExtensionBuilder<HKObje
                     case HKMetadataKeyWeatherHumidity:
                         extensionValue = .quantity(value.buildQuantity(mapping: .weatherHumidity))
                     case HKMetadataKeySessionEstimate:
-                        guard let sample = sample as? HKQuantitySample,
+                        guard let sample = object as? HKQuantitySample,
                               let mapping = HKQuantitySampleMapping.default[sample.quantityType] else {
                             continue // should be unreachable. skipping
                         }
@@ -205,15 +108,6 @@ extension FHIRExtensionBuilderProtocol where Self == FHIRExtensionBuilder<HKObje
                 observation.appendExtension(metadataExtension, replaceAllExistingWithSameUrl: true)
             }
         }
-    }
-}
-
-
-extension HKSourceRevision {
-    /// We define this as an optional String objc-compatible property, so that we can encode it into an Extension using the API we have above.
-    @objc fileprivate var OSVersion: String? {
-        let version = operatingSystemVersion
-        return "\(version.majorVersion).\(version.minorVersion).\(version.patchVersion)"
     }
 }
